address_old/app/requirements.txt
Django==3.2
djangorestframework==3.12.2
django-cors-headers==3.7.0
requests==2.25.1
